requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
html2text==2020.1.16
openai==1.12.0
pandas>=1.5.0
openpyxl>=3.0.0
Pillow>=8.0.0
rembg>=2.0.0
numpy>=1.19.0
opencv-python>=4.5.0
onnxruntime==1.15.1
scipy>=1.7.0
PyInstaller>=5.0.0 